+++ lib/src/data_source/nd_json_file.rs
@@ -1,4 +1,4 @@
-// raw_json_file.rs - Raw JSON file to RecordBatch
+// nd_json_file.rs - ndJSON file to RecordBatch
 // Sasaki, Naoki <nsasaki@sal.co.jp> March 25, 2023
 //
 
@@ -23,7 +23,8 @@ pub fn to_record_batch(
     let builder = json::ReaderBuilder::new(SchemaRef::new(if let Some(schema) = schema {
         schema.to_datafusion_schema()
     } else {
-        let (inferred_schema, _) = infer_json_schema_from_seekable(&mut file_reader, options.infer_schema_rows)?;
+        let (inferred_schema, _) =
+            infer_json_schema_from_seekable(&mut file_reader, options.infer_schema_rows)?;
         inferred_schema
     }));